Listing 10-1 Basic Function; No Return
The program displays a prompt five times, allowing you to type various com-
mands. Of course, nothing happens when you type, although you can pro-
gram those actions later, if you like. Here’s how this program works in regard
to creating a function:
Line 3 lists the function prototype. It’s essentially a copy of the first line of
the function (from Line 22), but ending with a semicolon. It can also be writ-
ten like this:
void prompt(void);
Because the function doesn’t require any arguments (the items in parentheses),
you can use the void keyword in there as well.
Line 13 accesses the function. The function is called as its own statement. It
doesn’t require any arguments or return any values, and it appears on a line
by itself, as shown in the listing. When the program encounters that state-
ment, program execution jumps up to the function. The function’s statements are executed, and then control returns to the next line in the code after the
function was called.
Lines 22 through 25 define the function itself. The function type is specified
on Line 22, followed by the function name, and then the parentheses. As with
the prototype, you can specify void in the parentheses because no argument
is passed to the function.
The function’s sole statement is held between curly brackets. The prompt()
function merely outputs a prompt by using the printf() function, which
makes it seem like the function isn’t necessary, but many examples of one-
line functions can be found in lots of programs